# Conflicts:
#	converage/BankModel.php.html
#	converage/BaseModel.php.html
#	converage/FactoryPattern.php.html
#	converage/UserModel.php.html
#	converage/dashboard.html
#	converage/index.html
#	list_bank.php
#	models/BankModel.php
#	models/FactoryPattern.php
#	models/UserModel.php
#	phpunit.xml
#	tests/UserModelTest.php

+++ models/UserModel.php
@@ -25,9 +25,14 @@ class UserModel extends BaseModel {
      * @return array
      */
     public function auth($userName, $password) {
-        $md5Password = md5($password);
-        $sql = 'SELECT * FROM users WHERE name = "' . $userName . '" AND password = "'.$md5Password.'"';
-        $user = $this->select($sql);
+        if(is_string($userName) || is_string($password)){
+            $md5Password = md5($password);
+            $sql = 'SELECT * FROM users WHERE name = "' . $userName . '" AND password = "'.$md5Password.'"';
+            $user = $this->select($sql);
+        }
+        else{
+            $user = [];
+        }
         return $user;
     }
 
@@ -37,10 +42,19 @@ class UserModel extends BaseModel {
      * @return mixed
      */
     public function deleteUserById($id) {
-      
-        $sql = 'DELETE FROM users WHERE id = '.$id;
-        return $this->delete($sql);
-
+        if(is_numeric($id)){    
+            if(is_float($id)){
+                return false;
+            }   
+            else{
+                $sql = 'DELETE FROM users WHERE id = '.$id;
+                return $this->delete($sql);
+            }   
+           
+        } 
+        else{
+            return false;
+        }
     }
 
 
@@ -50,15 +64,33 @@ class UserModel extends BaseModel {
      * @param $input
      * @return mixed
      */
-    public function updateUser($input) {
-        $sql = 'UPDATE users SET 
-                 name = "' . mysqli_real_escape_string(self::$_connection, $input['name']) .'", 
-                 password="'. md5($input['password']) .'"
-                WHERE id = ' . $input['id'];
-
-        $user = $this->update($sql);
-      
-        return $user;
+    public function updateUser($input, $bankModel) { 
+        if(isset($input['user_id'])){
+            $bankModel->updateBank($input);
+        }
+        else{
+            $t = base64_decode($input['version']);
+            $str = substr($t,18);
+            $temp = 'SELECT version FROM users WHERE id = '.$input['id'].'';
+            $newTemp = $this->select($temp);
+            if($newTemp[0]['version'] == $str){
+                $newV = $str+1;
+                $sql = 'UPDATE users SET 
+                    name = "' . $input['name'] .'", 
+                    email = "'.$input['email'].'",
+                    fullname = "'.$input['fullname'].'",
+                    password="'. md5($input['password']) .'", type = "'.$input['type'].'", version = "'.$newV.'"
+                    WHERE id = ' . $input['id'] ;
+                $user = $this->update($sql);  
+                //header('location: list_users.php?success');
+                return $user;           
+            } 
+            else{                
+                //header('location: list_users.php?err');  
+                return false;
+            }    
+        }
+           
     }
 
     /**
@@ -66,18 +98,24 @@ class UserModel extends BaseModel {
      * @param $input
      * @return mixed
      */
-    public function insertUser($input) {
-        $sql = "INSERT INTO `app_web1`.`users` (`name`, `password`,`fullname`,`email`,`type`) VALUES (" .
-        "'" . $input['name'] . "', '"
-        . md5($input['password']) . "', '"
-        . $input['fullname'] . "', '"
-        . $input['email'] . "', '"
-        . $input['type']
-        . "')";
-        $user = $this->insert($sql);
-      
-        return $user;
-                
+    public function insertUser($input, $bankModel) {
+        if(isset($input['user_id'])){
+            $bankModel->insertBank($input);
+        }
+        else{
+            $sql = "INSERT INTO `app_web1`.`users` (`name`, `password`,`fullname`,`email`,`type`,`version`) VALUES (" .
+            "'" . $input['name'] . "', '"
+            . md5($input['password']) . "', '"
+            . $input['fullname'] . "', '"
+            . $input['email'] . "', '"
+            . $input['type']
+            . "', '"
+            . 1
+            . "')";   
+            $user = $this->insert($sql);
+            return $user;
+        }      
+                 
     }
 
     /**
@@ -89,17 +127,20 @@ class UserModel extends BaseModel {
         //Keyword
         if (!empty($params['keyword'])) {
             $sql = 'SELECT * FROM users WHERE name LIKE "%' . $params['keyword'] .'%"';
-
+           
             //Keep this line to use Sql Injection
             //Don't change
             //Example keyword: abcef%";TRUNCATE banks;##
-            $users = self::$_connection->multi_query($sql);
+            $users = $this->select($sql);
+            // $users = self::$_connection->multi_query($sql);
+            
         } else {
+           
             $sql = 'SELECT * FROM users';
             $users = $this->select($sql);
         }
-
         return $users;
+       
     }
 
     public static function getInstance() {
@@ -114,7 +155,4 @@ class UserModel extends BaseModel {
      * @param $a
      * @param $b
      */
-    public function sumb($a, $b) {
-        return $a + $b;
-    }
-}   
+}
